remove deadlines-api; move classes to deadlines package
