+++ deadlines/build.gradle
@@ -3,7 +3,6 @@ apply plugin: 'com.palantir.metric-schema'
 apply plugin: 'com.palantir.revapi'
 
 dependencies {
-    api project(':deadlines-api')
     implementation 'com.palantir.tracing:tracing'
     implementation 'io.dropwizard.metrics:metrics-core'
     implementation 'com.palantir.tritium:tritium-metrics'
