+++ deadlines/src/main/java/com/palantir/deadlines/DeadlinesHttpHeaders.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.deadlines.api;
+package com.palantir.deadlines;
 
 public interface DeadlinesHttpHeaders {
     String EXPECT_WITHIN = "Expect-Within";
