+++ settings.gradle
@@ -11,4 +11,3 @@ apply plugin: 'com.palantir.jdks.settings'
 rootProject.name = 'deadlines-java-root'
 
 include 'deadlines'
-include 'deadlines-api'
